sui-indexer-alt-framework: migrate MockStore imports to trait crate
