+++ crates/sui-indexer-alt-framework/src/pipeline/concurrent/committer.rs
@@ -224,12 +224,14 @@ mod tests {
 
     use anyhow::ensure;
     use async_trait::async_trait;
+    use sui_indexer_alt_framework_store_traits::testing::mock_store::ConnectionFailure;
+    use sui_indexer_alt_framework_store_traits::testing::mock_store::MockConnection;
+    use sui_indexer_alt_framework_store_traits::testing::mock_store::MockStore;
     use sui_types::full_checkpoint_content::Checkpoint;
     use tokio::sync::mpsc;
 
     use crate::FieldCount;
     use crate::metrics::IndexerMetrics;
-    use crate::mocks::store::*;
     use crate::pipeline::Processor;
     use crate::pipeline::WatermarkPart;
     use crate::pipeline::concurrent::BatchStatus;
