+++ crates/sui-indexer-alt-framework/src/pipeline/concurrent/reader_watermark.rs
@@ -104,12 +104,14 @@ mod tests {
     use std::sync::Arc;
 
     use async_trait::async_trait;
+    use sui_indexer_alt_framework_store_traits::testing::mock_store::MockConnection;
+    use sui_indexer_alt_framework_store_traits::testing::mock_store::MockStore;
+    use sui_indexer_alt_framework_store_traits::testing::mock_store::MockWatermark;
     use sui_pg_db::FieldCount;
     use sui_types::full_checkpoint_content::Checkpoint;
     use tokio::time::Duration;
 
     use crate::metrics::IndexerMetrics;
-    use crate::mocks::store::*;
     use crate::pipeline::Processor;
     use crate::pipeline::concurrent::BatchStatus;
